cleanup includes removal of duplicate ids
when a child (any user actually) tries the same book that hasn't been downloaded several times, its id is added several times to the list and the script will download the same book several times. This modification removes duplicate ids in the cleaning process, so the update process runs a lot faster.